requirements.txt
Django==3.1.4
django-admin-interface==0.14.2
django-apptemplates==1.5
django-colorfield==0.3.2
django-extensions==3.1.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-pipeline==2.0.5
django-treebeard==4.3.1
psycopg2-binary==2.8.6
toolz==0.11.1
uWSGI==2.0.19.1